requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangChain and AI dependencies
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-aws==0.1.0
langchain-huggingface==0.0.1
langgraph==0.0.26

# Vector store and embeddings
faiss-cpu==1.7.4

# Text processing
langchain-text-splitters==0.0.1

# Environment and utilities
python-dotenv==1.0.0
boto3==1.34.0
botocore==1.34.0

# Additional utilities
requests==2.31.0
numpy==1.24.3
